requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
black=21.12b0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bottleneck=1.3.2=py39h7cc1a96_1
brotli=1.0.9=ha925a31_2
ca-certificates=2021.10.26=haa95532_2
certifi=2021.10.8=py39haa95532_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.4=pypi_0
click=8.0.1=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pypi_0
configparser=5.0.2=pypi_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py39haa95532_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=pypi_0
executing=0.8.2=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
gym=0.21.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.2=pypi_0
intel-openmp=2021.3.0=haa95532_3372
ipykernel=6.7.0=pypi_0
ipython=8.0.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.1=py39hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.17.2=h23ce68f_1
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.0.1=pypi_0
matplotlib=3.4.2=py39haa95532_0
matplotlib-base=3.4.2=py39h49ac443_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.0=py39h277e83a_2
mkl_random=1.2.2=py39hf11a4ad_0
mpl-axes-aligner=1.3=pypi_0
munkres=1.1.4=py_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.10=pypi_0
nbconvert=6.4.0=pypi_0
nbformat=5.1.3=pypi_0
nest-asyncio=1.5.4=pypi_0
ninja=1.10.2=h6d14046_1
notebook=6.4.7=pypi_0
numexpr=2.7.3=py39hb80d3ca_1
numpy=1.21.1=pypi_0
numpy-base=1.20.3=py39hc2deb75_0
olefile=0.46=py_0
openssl=1.1.1l=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.3.1=py39h6214cd6_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
pathtools=0.1.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py39h4fa10fc_0
pip=21.2.3=py38haa95532_0
platformdirs=2.4.1=pypi_0
prometheus-client=0.12.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.24=pypi_0
protobuf=3.17.3=pypi_0
psutil=5.8.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pypi_0
pygments=2.11.2=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.18.1=pypi_0
python=3.9.6=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.9.0=py3.9_cuda10.2_cudnn7_0
pytz=2021.1=pyhd3eb1b0_0
pywin32=303=pypi_0
pywinpty=1.1.6=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
qt=5.9.7=vc14h73c81de_0
qtconsole=5.2.2=pypi_0
qtpy=2.0.0=pypi_0
requests=2.26.0=pypi_0
scipy=1.7.1=py39hbe87c03_2
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.3.1=pypi_0
setuptools=52.0.0=py39haa95532_0
shortuuid=1.0.1=pypi_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pypi_0
sqlite=3.36.0=h2bbff1b_0
stack-data=0.1.4=pypi_0
subprocess32=3.5.4=pypi_0
tensorboardx=2.2=pyhd3eb1b0_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
tk=8.6.10=he774522_0
tomli=1.2.3=pypi_0
torchvision=0.10.0=pypi_0
tornado=6.1=py39h2bbff1b_0
tqdm=4.62.0=pyhd3eb1b0_1
traitlets=5.1.1=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.12.10=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wincertstore=0.2=py39h2bbff1b_0
xz=5.2.5=h62dcd97_0
yaspin=2.1.0=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
